+++ packages/components/inputs/money-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/money-input
 
+## 13.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/select-utils@13.0.1
+
 ## 13.0.0
 
 ### Patch Changes