+++ packages/components/primary-action-dropdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/primary-action-dropdown
 
+## 13.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/buttons@13.0.1
+
 ## 13.0.0
 
 ### Patch Changes